refactor(ui): Tooltip - convert to typescript
