+++ packages/ui/src/ui/tooltip/tooltip.stories.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Story } from '@storybook/react'; // eslint-disable-line
 
 import { getWrapperDecorator } from '../../stories';
 import { Tooltip } from '.';
@@ -12,17 +13,17 @@ export default {
   },
 };
 
-const Template = (props) => <Tooltip {...props}>Job #123</Tooltip>;
+const Template = (props: any) => <Tooltip {...props}>Job #123</Tooltip>;
 
-export const Default = Template.bind();
+export const Default: Story = Template.bind({});
 
-export const LongTooltip = Template.bind();
+export const LongTooltip: Story = Template.bind({});
 
 LongTooltip.args = {
   title: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam molestie neque id lectus mollis, et imperdiet libero porta.',
 };
 
-export const DarkModeFalse = Template.bind();
+export const DarkModeFalse: Story = Template.bind({});
 
 DarkModeFalse.args = {
   darkMode: false,
